\name{respiratory}
\docType{data}
\alias{respiratory}
\title{Removal times by age for a respiratory epidemic that occurred on the island of Tristan da Cunha, 1967}
\description{
    Removal times and removals by 'age'.
}
\usage{data(respiratory)}
\format{A \code{list} containing 2 elements, a \code{vector} with removal data (respiratoryRemovals) and a matrix with removals by 'age' and day of removal (removals.by.day).}
\references{
Becker, N.G. & Hopper, J.L. (1983). 'Assessing heterogeneity of disease spread through a community' \emph{American J. Epidemiology} \bold{117}, 362-374.

}
\keyword{datasets}
