\name{smallpox}
\docType{data}
\alias{smallpox}
\title{Removal times for smallpox data}
\description{
    Removal times associated with an outbreak of smallpox in a closed community of 120 individuals in Abakaliki, Nigeria.
}
\usage{data(smallpox)}
\format{A \code{vector} containing 30 observations}
\references{
    Bailey, N.T.J. (1975). \emph{The Mathematical Theory of Infectious Diseases and its Applications}, 2nd edn. London: Griffin
}
\keyword{datasets}
